msm: mdss: clean any dirty pipes after a successful prepare
After prepare call is made it should be assumed that all pipes that are
needed for this round should already be set. If there are any pipes that
are still dirty after a successful prepare call, it means that they are
not properly referenced in user land. Thus mark them for cleanup.

Change-Id: I0987ab4c4cf7ca619a7363f995ef11d371a1f2f2
Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: placiano <placiano80@gmail.com>