msm: mdss: change the print level to debug for dynamic fps write
During stability testing, the device can enter and exit the idle
screen use case very frequently. When this happens the fps switch
between the minimum and maximum configured values happens many
times causing the kernel logs to be flooded with messages related
to this. Hence, silence this print statement by changing it to
pr_debug.

Change-Id: I06a3b6560cc44d94202c946aec78a0432342018a
Signed-off-by: Padmanabhan Komanduru <pkomandu@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: placiano <placiano80@gmail.com>